+++ frontend/src/pages/Console/Console.css
@@ -1,4 +1,50 @@
+.console {
+    height: 100%;
+    width: 60%;
+    font-family: Consolas,monaco,monospace; 
+}
+
+.console-output {
+    color: white;
+    padding: 20px;
+    height: 75vh;
+    width: 100%;
+    overflow: hidden;
+    overflow-y: scroll;
+    background-color: rgb(70, 70, 70);
+}
+
+.console input {
+    color: white;
+    width: 100%;
+    height: 30px;
+    background-color: rgb(70, 70, 70);
+}
+
+input[type="text"],textarea, select { 
+    outline: none;
+}
+
+#style-2::-webkit-scrollbar-track
+{
+	-webkit-box-shadow: inset 0 0 6px rgba(0,0,0,0.3);
+	border-radius: 10px;
+	background-color: #555454;
+}
+
+#style-2::-webkit-scrollbar
+{
+	width: 12px;
+	background-color: rgb(73, 68, 68);
+}
+
+#style-2::-webkit-scrollbar-thumb
+{
+	border-radius: 10px;
+	-webkit-box-shadow: inset 0 0 6px rgba(0,0,0,.3);
+	background-color: #8d8c8c;
+}
+
 #clear-btn {
     margin-top: 20px;
-
 }
