+++ frontend/src/pages/Console/ServerConsole.css
